Move Problem into its own file.
Because I am tired of forgetting where it is.
