+++ lint/response.go
@@ -14,38 +14,8 @@
 
 package lint
 
-import "github.com/jhump/protoreflect/desc"
-
 // Response describes the result returned by a rule.
 type Response struct {
 	FilePath string    `json:"file_path" yaml:"file_path"`
 	Problems []Problem `json:"problems" yaml:"problems"`
-}
-
-// Problem contains information about a result produced by an API Linter.
-type Problem struct {
-	// Message provides a short description of the problem.
-	Message string `json:"message" yaml:"message"`
-
-	// Suggestion provides a suggested fix, if applicable.
-	Suggestion string `json:"suggestion,omitempty" yaml:"suggestion,omitempty"`
-
-	// Location provides the location of the problem. If both
-	// `Location` and `Descriptor` are specified, the location
-	// is then used from `Location` instead of `Descriptor`.
-	Location Location `json:"location" yaml:"location"`
-
-	// Descriptor provides the descriptor related
-	// to the problem. If present and `Location` is not
-	// specified, then the starting location of the descriptor
-	// is used as the location of the problem.
-	Descriptor desc.Descriptor `json:"-" yaml:"-"`
-
-	// RuleID provides the ID of the rule that this problem belongs to.
-	// DO NOT SET: this field will be set by the linter based on rule info
-	// and user configs.
-	RuleID RuleName `json:"rule_id" yaml:"rule_id"`
-
-	// DO NOT SET:  this field will be set by the linter based on user configs.
-	Category string `json:"category,omitempty" yaml:"category,omitempty"`
 }
